Disentangle remote execution from permission management
This allows having a permission manager without having messaging
and remote execution, which will be useful in the almond-enterprise
use case, and for multi-user almond in general
