+++ lib/engine.js
@@ -28,6 +28,13 @@ const Config = require('./config');
 const sqlite = require('./db/sqlite');
 const Memory = require('./db/memory');
 
+// FINISHME
+class DummyGroupDelegate {
+    getGroups(principal) {
+        return Promise.resolve([]);
+    }
+}
+
 module.exports = class Engine {
     constructor(platform, options = {}) {
         // constructor
@@ -40,6 +47,7 @@ module.exports = class Engine {
         const hasMemory = platform.hasFeature('memory');
         const hasDiscovery = platform.hasFeature('discovery');
         const hasPermissions = platform.hasFeature('permissions');
+        const hasRemote = platform.hasFeature('remote');
 
         // tiers and devices are always enabled
         this._tiers = new TierManager(platform);
@@ -57,7 +65,7 @@ module.exports = class Engine {
         else
             this._memory = null;
 
-        if (hasApps)
+        if (hasApps || hasPermissions)
             this._schemas = new ThingTalk.SchemaRetriever(this._thingpedia, this._memory);
         else
             this._schemas = null;
@@ -99,17 +107,25 @@ module.exports = class Engine {
             this._discovery = null;
         }
         if (hasPermissions) {
-            if (!hasMessaging || !hasApps)
-                throw new TypeError('Permission manager requires messaging and apps');
             const PermissionManager = require('./permissions/permission_manager');
-            const CommunicationManager = require('./messaging/communication_manager');
 
-            this._permissionManager = new PermissionManager(this._platform, this._messaging, this._schemas);
+            let groupDelegate;
+            if (platform.hasCapability('permission-groups'))
+                groupDelegate = platform.getCapability('permission-groups');
+            else
+                groupDelegate = new DummyGroupDelegate();
+            this._permissionManager = new PermissionManager(this._platform, groupDelegate, this._schemas);
             this._modules.push(this._permissionManager);
+        } else {
+            this._permissionManager = null;
+        }
+        if (hasRemote) {
+            if (!hasMessaging || !hasPermissions || !hasApps)
+                throw new TypeError('Remote execution requires messaging, permission management and apps');
+            const CommunicationManager = require('./messaging/communication_manager');
             this._remote = new CommunicationManager(this._platform, this._permissionManager, this._messaging, this._tiers, this._devices, this._schemas);
             this._modules.push(this._remote);
         } else {
-            this._permissionManager = null;
             this._remote = null;
         }
