+++ lib/messaging/communication_manager.js
@@ -15,6 +15,7 @@ const ThingTalk = require('thingtalk');
 const Protocol = require('../tiers/protocol');
 const ArraySet = require('../util/array_set');
 const { ChannelStateBinder } = require('../db/channel');
+const RemoteProgramExecutor = require('./remote_program_executor');
 
 const CURRENT_VERSION = 3;
 
@@ -249,9 +250,9 @@ class SharedProgramState {
 }
 
 module.exports = class CommunicationManager {
-    constructor(platform, executor, messaging, tierManager, devices, schemaRetriever) {
+    constructor(platform, permissions, messaging, tierManager, devices, schemaRetriever) {
         this._platform = platform;
-        this._executor = executor;
+        this._executor = new RemoteProgramExecutor(platform, messaging, permissions);
         this._messaging = messaging;
         this._tierManager = tierManager;
         this._devices = devices;
