+++ lib/permissions/permission_manager.js
@@ -14,23 +14,6 @@ const uuid = require('uuid');
 const ThingTalk = require('thingtalk');
 const PermissionSql = require('../db/permissions');
 
-function throwCode(code, message) {
-    var e = new Error(message);
-    e.code = code;
-    return e;
-}
-
-class GroupDelegate {
-    constructor(messaging) {
-        this._messaging = messaging;
-    }
-
-    getGroups(principal) {
-        // FINISHME
-        return Promise.resolve([]);
-    }
-}
-
 function getCompatKey(permissionRule) {
     let key = '';
     if (permissionRule.query.isStar)
@@ -50,15 +33,14 @@ function getCompatKey(permissionRule) {
 }
 
 module.exports = class PermissionManager {
-    constructor(platform, messaging, schemaRetriever) {
+    constructor(platform, groupDelegate, schemaRetriever) {
         this._platform = platform;
-        this._messaging = messaging;
         this._db = new PermissionSql(platform);
 
         this._permissiondb = new Map;
         this._permissionextra = new Map;
         this._checker = new ThingTalk.PermissionChecker(platform.getCapability('smt-solver'),
-            schemaRetriever, new GroupDelegate(messaging));
+            schemaRetriever, groupDelegate);
     }
 
     start() {
@@ -120,30 +102,6 @@ module.exports = class PermissionManager {
         return this._db.deleteOne(uniqueId);
     }
 
-    _askForPermission(principal, identity, program) {
-        var assistant = this._platform.getCapability('assistant');
-        if (!assistant)
-            throw throwCode('ENOTSUP', 'Interaction not supported');
-        var conversation = assistant.getConversation();
-        if (!conversation)
-            throw throwCode('EPERM', 'User not available to confirm');
-
-        return conversation.askForPermission(principal, identity, program);
-    }
-
-    _verifyIdentity(principal, identity) {
-        if (!identity)
-            return Promise.reject(throwCode('EINVAL', 'Invalid identity'));
-        if (!identity.startsWith('phone:') && !identity.startsWith('email:') && !identity.startsWith('omlet:') && !identity.startsWith('matrix-account:'))
-            return Promise.reject(throwCode('EINVAL', 'Invalid identity ' + identity));
-        return this._messaging.getAccountForIdentity(identity).then((account) => {
-            if (!account)
-                throw throwCode('EINVAL', 'Invalid identity ' + identity);
-            if (this._messaging.type + '-account:' + account !== principal)
-                throw throwCode('EPERM', 'Identity does not match principal');
-        });
-    }
-
     checkCanBeAllowed(principal, program) {
         return this._checker.check(ThingTalk.Ast.Value.Entity(principal, 'tt:contact', null), program,
                                    { transform: false });
@@ -152,54 +110,5 @@ module.exports = class PermissionManager {
     checkIsAllowed(principal, program) {
         return this._checker.check(ThingTalk.Ast.Value.Entity(principal, 'tt:contact', null), program,
                                    { transform: true });
-    }
-
-    _isAllowedProgram(principal, identity, program) {
-        return this._verifyIdentity(principal, identity).then(() => {
-            return this.checkIsAllowed(principal, program);
-        }).then((newProgram) => {
-            if (newProgram !== null) {
-                return newProgram;
-            } else {
-                // ask the user for a new policy
-                return this._askForPermission(principal, identity, program);
-            }
-        });
-    }
-
-    _cleanProgram(program) {
-        for (let [,slot,,] of program.iterateSlots()) {
-            if (slot instanceof ThingTalk.Ast.Selector) {
-                slot.principal = null;
-            } else if (slot.value.isEntity) {
-                slot.value.display = null;
-            } else if (slot.value.isLocation) {
-                if (slot.value.isAbsolute)
-                    slot.value.display = null;
-                else
-                    throwCode('EPERM', 'Relative locations are not allowed');
-            }
-        }
-    }
-
-    installProgram(principal, identity, program, uniqueId) {
-        if (program.principal !== null)
-            return Promise.reject(throwCode('EPERM', 'Permission denied'));
-        return Promise.resolve().then(() => {
-            this._cleanProgram(program);
-            return this._isAllowedProgram(principal, identity, program);
-        }).then((newProgram) => {
-            if (newProgram === null)
-                throw throwCode('EPERM', 'Permission denied');
-
-            var assistant = this._platform.getCapability('assistant');
-            if (!assistant)
-                throw throwCode('ENOTSUP', 'Interaction not supported');
-            var conversation = assistant.getConversation();
-            if (!conversation)
-                throw throwCode('EPERM', 'User not available to confirm');
-
-            return conversation.runProgram(newProgram, uniqueId, identity);
-        }).then(() => undefined);
     }
 };
