+++ test/index.js
@@ -42,7 +42,7 @@ const THINGPEDIA_URL = 'https://almond-dev.stanford.edu/thingpedia';
 
 async function runTests(engine, limitTo) {
     try {
-        for (let x of ['devices', 'apps', 'http_client', 'util', 'builtins']) {
+        for (let x of ['devices', 'apps', 'http_client', 'util', 'builtins', 'permissions']) {
             if (limitTo !== undefined && x !== limitTo)
                 continue;
             console.log(`Running ${x} tests`);
